--- src/main/java/com/fires/fires/asset/contant/AssetCategory_back.java
@@ -1,10 +1,0 @@
-package com.fires.fires.assets.contant;
-
-/**
- * @deprecated (MarketType 재정의)
- */
-@Deprecated(since = "1")
-public enum AssetCategory_back {
-    //todo: category1 : domestic, overseas category2 : stock, etf, etn, category3 : markey
-    ALL, OVERSEAS_STOCK, DOMESTIC_STOCK, OVERSEAS_ETF, DOMESTIC_ETF
-}
