+++ src/main/java/com/fires/fires/asset/contant/Country.java
@@ -1,4 +1,4 @@
-package com.fires.fires.assets.contant;
+package com.fires.fires.asset.contant;
 
 public enum Country {
     KOR, USA
