+++ src/main/java/com/fires/fires/asset/dto/AssetDto.java
@@ -1,8 +1,8 @@
-package com.fires.fires.assets.asset.dto;
+package com.fires.fires.asset.dto;
 
 
-import com.fires.fires.assets.asset.entity.Asset;
-import com.fires.fires.assets.asset.entity.embed.Category;
+import com.fires.fires.asset.entity.Asset;
+import com.fires.fires.asset.entity.embed.Category;
 
 public record AssetDto(Long assetId,
                        String symbol,
