+++ src/main/java/com/fires/fires/asset/entity/Asset.java
@@ -1,7 +1,7 @@
-package com.fires.fires.assets.asset.entity;
+package com.fires.fires.asset.entity;
 
-import com.fires.fires.assets.asset.entity.embed.Category;
-import com.fires.fires.assets.contant.MarketType;
+import com.fires.fires.asset.entity.embed.Category;
+import com.fires.fires.asset.contant.MarketType;
 import com.fires.fires.common.entity.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
