+++ src/main/java/com/fires/fires/asset/repository/AssetRepository.java
@@ -1,6 +1,6 @@
-package com.fires.fires.assets.asset.repository;
+package com.fires.fires.asset.repository;
 
-import com.fires.fires.assets.asset.entity.Asset;
+import com.fires.fires.asset.entity.Asset;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
