+++ src/main/java/com/fires/fires/external/exception/KisException.java
@@ -1,11 +1,12 @@
-package com.fires.fires.common.exception;
+package com.fires.fires.external.exception;
 
+import com.fires.fires.common.exception.BaseException;
 import org.springframework.http.HttpStatus;
 
 /**
  * 한국 투자증권 api호출 중 발생한 exception처리용
  */
-public class KisException extends BaseException{
+public class KisException extends BaseException {
     public KisException(HttpStatus httpStatus, Throwable e) {
         super(httpStatus, e);
     }
