+++ src/main/java/com/fires/fires/member/controller/MemberController.java
@@ -3,8 +3,9 @@ package com.fires.fires.member.controller;
 import com.fires.fires.common.constant.ResponseCode;
 import com.fires.fires.common.dto.Response;
 import com.fires.fires.member.dto.MemberDto;
-import com.fires.fires.member.dto.MemberSignupRequest;
+import com.fires.fires.member.dto.request.MemberSignupRequest;
 import com.fires.fires.member.service.MemberWriteService;
+import com.fires.fires.member.service.MemberWriteServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
