+++ src/main/java/com/fires/fires/member/dto/request/MemberSignupRequest.java
@@ -1,4 +1,4 @@
-package com.fires.fires.member.dto;
+package com.fires.fires.member.dto.request;
 
 import com.fires.fires.member.constant.Provider;
