+++ src/main/java/com/fires/fires/portfolio/constant/Currency.java
@@ -1,7 +1,7 @@
-package com.fires.fires.common.constant;
+package com.fires.fires.portfolio.constant;
 
 /**
- * 통화정보, 애플리케이션 전반에서 활용
+ * 통화정보
  */
 public enum Currency {
     KRW, USD
