+++ src/main/java/com/fires/fires/portfolio/dto/AssetInfoDto.java
@@ -1,6 +1,6 @@
 package com.fires.fires.portfolio.dto;
 
-import com.fires.fires.assets.asset.entity.Asset;
+import com.fires.fires.asset.entity.Asset;
 
 public record AssetInfoDto(Asset asset,
                            Long assetNumber,
